move staking mutation to staking component
